+++ MatrixKit/Models/Room/MXKRoomDataSourceManager.h
@@ -20,7 +20,7 @@
 
 /**
  `MXKRoomDataSourceManagerReleasePolicy` defines how a `MXKRoomDataSource` instance must be released
- when [MXKRoomDataSourceManager closeRoomDataSource:] is called.
+ when [MXKRoomDataSourceManager closeRoomDataSourceWithRoomId:] is called.
  
  Once released, the in-memory data (messages that are outgoing, failed sending, ...) of room data source
  is lost.
@@ -86,9 +86,9 @@ typedef enum : NSUInteger {
  
  @param roomId the room id of the room.
  @param create if YES, the MXKRoomDataSourceManager will create the room data source if it does not exist yet.
- @return the room data source (instance of MXKRoomDataSource-inherited class)
+ @param onComplete blocked with the room data source (instance of MXKRoomDataSource-inherited class).
  */
-- (MXKRoomDataSource*)roomDataSourceForRoom:(NSString*)roomId create:(BOOL)create;
+- (void)roomDataSourceForRoom:(NSString*)roomId create:(BOOL)create onComplete:(void (^)(MXKRoomDataSource *roomDataSource))onComplete;
 
 /**
  Make a room data source be managed by the manager.
@@ -105,10 +105,10 @@ typedef enum : NSUInteger {
  
  The roomDataSource instance will be actually destroyed according to the current release policy.
 
- @param roomDataSource the data source to release.
- @param forceRelease if yes the roomDataSource instance will be destroyed whatever the policy is.
+ @param roomId the room if of the data source to release.
+ @param forceRelease if yes the room data source instance will be destroyed whatever the policy is.
  */
-- (void)closeRoomDataSource:(MXKRoomDataSource*)roomDataSource forceClose:(BOOL)forceRelease;
+- (void)closeRoomDataSourceWithRoomId:(NSString*)roomId forceClose:(BOOL)forceRelease;
 
 /**
  The release policy to apply when `MXKRoomDataSource` instances are closed.