+++ MatrixKit/Views/PushRule/MXKPushRuleTableViewCell.m
@@ -69,7 +69,7 @@
             MXRoom *room = [_mxSession roomWithRoomId:mxPushRule.ruleId];
             if (room)
             {
-                description = [NSString stringWithFormat:[NSBundle mxk_localizedStringForKey:@"notification_settings_room_rule_title"], room.state.displayname];
+                description = [NSString stringWithFormat:[NSBundle mxk_localizedStringForKey:@"notification_settings_room_rule_title"], room.summary.displayname];
             }
             break;
         }