+++ CHANGES.rst
@@ -1,3 +1,21 @@
+Changes in MatrixKit in 0.8.0 (2018-08-10)
+==========================================
+
+Improvements:
+ * Upgrade MatrixSDK version (v0.11.0).
+ * MXKRoomDataSource: Add send reply with text message (vector-im/riot-ios#1911).
+ * MXKSessionRecentsDataSource: Hide rooms that should not be displayed to user (linked to vector-im/riot-ios#1938).
+ * MXKRoomDataSource: Fix a multithreading issue that caused a crash (PR #456).
+ 
+Bug fix:
+ * MXKSampleJSQMessagesViewController: Fix room display assertion when user has no display name.
+
+API break:
+ * MXKContactManager: Remove the privateMatrixContacts method.
+ * MXKSearchCellDataStoring: Replace initWithSearchResult by async cellDataWithSearchResult.
+ * MXKRoomDataSourceManager: The roomDataSourceForRoom method is now asynchronous.
+ * MXKRoomDataSourceManager: closeRoomDataSource has been replaced by closeRoomDataSourceWithRoomId.
+
 Changes in MatrixKit in 0.7.15 (2018-07-03)
 ==========================================